third_party/blink/web_tests/platform/generic/fast/frames/adopt-from-created-document-expected.txt

ALERT: 1
ALERT: 2
ALERT: 3
ALERT: 4
ALERT: 5
ALERT: 6
ALERT: 7
